Rename 'engines' package to 'engine'
