+++ evalml/automl/__init__.py
@@ -1,3 +1,3 @@
 from .automl_search import AutoMLSearch
 from .utils import get_default_primary_search_objective, make_data_splitter, tune_binary_threshold
-from .engines import SequentialEngine, EngineBase
+from .engine import SequentialEngine, EngineBase
