+++ evalml/automl/automl_search.py
@@ -11,7 +11,7 @@ from .pipeline_search_plots import PipelineSearchPlots
 
 from evalml.automl.automl_algorithm import IterativeAlgorithm
 from evalml.automl.callbacks import log_error_callback
-from evalml.automl.engines import SequentialEngine
+from evalml.automl.engine import SequentialEngine
 from evalml.automl.utils import (
     get_default_primary_search_objective,
     make_data_splitter,
